小更新

+++ Source/main/main.cpp
@@ -8,28 +8,60 @@
 
 using namespace std;
 
-/**返回每个字的值**/
-template <class T>
-int value(T *a,int n,T str){
-	for(int i=0;i<n;i++){
-		if(a[i]==str) return i+1;
-	}
-	return -1;
-}
 
+/**----------词法分析器----------**/
+map<int,Token> scan(FILE * fpin);
 /** 显示词法分析结果 **/
 void showTokens(map<int,Token> tokens);
 
-/**----------词法分析器----------**/
-map<int,Token> ana(FILE * fpin);
 /**----------语法分析器----------**/
 
+// 获取语句流
+
+
+//语法结构
+
+// 赋值
+// var = value;
+int VOLUATION_ID = 1;
+string VOLUATION[4] = {
+	string(VAR),
+	string(1,OPERATOR[6]),
+	CONST,string(1,SEPARATER[0])};
+
+// 定义函数
+// KEYWORD VARNAME (){
+int DEFINE_ID = 2;
+string DEFINE[] = {
+	string(KEY),
+	string(VAR),
+	string(1,SEPARATER[6]),
+	string(1,SEPARATER[7]),
+	string(1,SEPARATER[2])};
+
+//获取语句
+/*
+map<int,string[]> get_sentence(map<int,Token> m)
+{
+	map<int,Token>::iterator it;
+	Token now;
+	for(it = m.begin();it!=m.end();it++){
+		now = it->second;
+		if(now.source != ";"){
+		}
+	}
+	return m;
+}
+*/
+
+//语法分析
+bool parsing(map<int,Token> m)
+{
+	return true;	
+}
 
 /**----------代码生成器----------**/
-string asm_print();
-string asm_string(string str);
-string asm_main();
-string asm_head();
+#include "asm.h"
 void generate(map<int,Token> m)
 {
 		
@@ -45,16 +77,30 @@ void generate(map<int,Token> m)
 
 	out_asm += asm_head();
 	out_asm += asm_main();
+	/*
+	out_asm += "\tstp x29, x30, [sp, #-16]!\n"
+    "\tadrp    x0, .content\n"
+    "\tadd x0, x0, :lo12:.content\n"
+	"\tmov x29, sp\n"
+    "\tbl  puts\n"
+	"\tmov w0, wzr\n"
+	"\tldp x29, x30, [sp], #16\n"
+	"\tret\n";
+	*/
 	map<int,Token>::iterator it;
 	it = m.begin();
 	while(it!=m.end()){
 		if(it->second.type==STRING){
-			out_asm += asm_string(it->second.source);
+			out_asm += asm_string("content",it->second.source);
+		}else if(it->second.type==VAR){
+			if(it->second.source=="print"){
+				/*
+				out_asm += asm_print(".content");
+				*/
+			}
 		}
 		it++;
 	}
-	//cout << "Finish"<<endl;
-
 	fprintf(out,"%s",out_asm.c_str());
 	// fputs(print_asm.c_str(), out);
 	fclose(out);
@@ -67,38 +113,49 @@ int main(int argc,char *argv[])
     char inFile[40];
     string analysed; 
     FILE *fpin;
-    //cout<<"File name:";
-    //cin>>inFile;
-    //fpin=fopen("src.cot","r");
-	
 
-	//参数分析
-	//fpin=fopen(argv[1],"r");
+	/** 参数分析 **/
+
 	int ASM_SOURCE_MODE = 0;
+	int SHOW_TOKENS = 0;
+	string outname = "tester";	//文件名
 	for(int i=1;i<argc;i++){
 		if(!strcmp(argv[i],"-S")){
 			ASM_SOURCE_MODE = 1;
+		}else if(!strcmp(argv[i],"--tokens")){
+			cout<<"词法分析"<<endl;
+			SHOW_TOKENS = 1;
+		}else if(!strcmp(argv[i],"-o")){
+			i++;
+			outname = argv[i];
+		}else if(!strcmp(argv[i],"-h") || !strcmp(argv[i],"--help")){
+			printf("%s\n",HELP_MSG);
+			return 0;
 		}else{
 			fpin=fopen(argv[i],"r");
-			break;
+			if(!fpin)
+				cout <<"请检查参数"<<endl;
 		}
 	}
-
-    //cout<<"1.词法分析"<<endl;
 	
-
-	
+	/** 词法分析 **/
     map<int,Token> tokens;
-    tokens = ana(fpin);
-    
-	/*
-    cout <<"Tokens----------------"<<endl;
-    showTokens(tokens);
-	*/
+    tokens = scan(fpin);
+	if(SHOW_TOKENS == 1){
+		showTokens(tokens);
+		return 1;
+	}
+
+	/** 语法分析 **/
+	//syntax(tokens);
+
+	/** 汇编代码生成 **/
+
 	generate(tokens);
 	
 	if(ASM_SOURCE_MODE == 0){
-		system("gcc test.s -o tester");
+		string compile ="gcc test.s -o "+outname;
+		system(compile.c_str());
 		system("rm test.s");
 	}
